pyproject.toml
[tool.poetry]
name = "WebZFS"
version = "0.55.0"
description = ""
authors = ["q5sys <jt@obs-sec.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
python-multipart = "^0.0.20"
uvicorn = "^0.34.2"
gunicorn = "^23.0.0"
typer = "^0.15.3"
pydantic-settings = "^2.9.1"
jinja2 = "^3.1.6"
psutil = "^7.0.0"
humanize = "^4.12.3"
python-jose = "^3.4.0"
python-pam = "^2.0.2"
paramiko = "3.5.0"
cryptography = "44.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.0"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
pre-commit = "^3.6.0"
pytest = "^8.3.0"
pytest-socket = "^0.6.0"
pytest-xdist = "^3.5.0"
pytest-watcher = "^0.3.4"
pytest-mock = "^3.12.0"
pytest-asyncio = "^0.24.0"
pytest-freezegun = "^0.4.2"
coverage = "^7.6.0"
httpx = "^0.28.0"
djhtml = "^3.0.6"
faker = "^22.5.1"
setuptools = "^75.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = 'black'
line_length = 89
skip = ['.venv']

[tool.black]
line-length = 89
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
    (
    \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    )
'''

[tool.ruff]
exclude = ["config/settings"]
line-length = 89
target-version = "py311"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_return_any = false
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_any_unimported = false
disallow_any_explicit = false
disallow_any_generics = true
show_error_codes = true
check_untyped_defs = true
exclude = ["config/settings", ".venv"]

[[tool.mypy.overrides]]
module = [
  "config.settings.*",
]
follow_imports = 'skip'

[[tool.mypy.overrides]]
module = [
  "views.*",
]
disallow_untyped_defs = false

[tool.coverage.run]
omit = ["config/settings/*"]

[tool.coverage.report]
sort = "Cover"
skip_covered = true
exclude_lines = [
  "pragma: no cover",
  "if TYPE_CHECKING:",
]

[tool.coverage.html]
skip_covered = true
